+++ link-grammar/Makefile.am
@@ -81,6 +81,7 @@ liblink_grammar_la_SOURCES =        \
 	api.c                            \
 	connectors.c                     \
 	dict-common/dict-common.c        \
+	dict-common/dict-impl.c          \
 	dict-common/dict-utils.c         \
 	dict-common/file-utils.c         \
 	dict-common/idiom.c              \
@@ -133,6 +134,7 @@ liblink_grammar_la_SOURCES =        \
 	dict-common/dict-api.h           \
 	dict-common/dict-common.h        \
 	dict-common/dict-defines.h       \
+	dict-common/dict-impl.h          \
 	dict-common/dict-structures.h    \
 	dict-common/dict-utils.h         \
 	dict-common/file-utils.h         \
