+++ link-grammar/dict-common/dict-impl.h
@@ -1,0 +1,7 @@
+
+#include "link-includes.h"
+
+// Already declared in link-includes.h
+// const char * linkgrammar_get_dict_locale(Dictionary dict);
+// const char * linkgrammar_get_version(void);
+// const char * linkgrammar_get_dict_version(Dictionary dict);
